+++ src/test/scala/org/podval/tools/test/testproject/GroupingFunSpec.scala
@@ -3,7 +3,7 @@ package org.podval.tools.test.testproject
 import org.gradle.api.internal.tasks.testing.junit.result.{TestClassResult, TestMethodResult}
 import org.gradle.api.tasks.testing.TestResult.ResultType
 import org.podval.tools.backend.BackendPlugin
-import org.podval.tools.build.{Dependency, ScalaBackend, ScalaBinaryVersion, ScalaVersion}
+import org.podval.tools.build.{Dependency, ScalaBackend, ScalaBinaryVersion, ScalaLibrary, ScalaVersion}
 import org.scalatest.funspec.AnyFunSpec
 import scala.jdk.CollectionConverters.*
 import ForClass.{ClassExpectation, MethodExpectation}
@@ -109,7 +109,7 @@ abstract class GroupingFunSpec extends AnyFunSpec:
     fixtures: Seq[Fixture],
     backend: ScalaBackend
   ): Seq[Fixture] = fixtures
-    .filter(_.framework.maker(backend).isDefined)
+    .filter(_.framework.forBackend(backend).isDefined)
 
   private def forBackends(
     projectName: Seq[String],
@@ -254,7 +254,11 @@ abstract class GroupingFunSpec extends AnyFunSpec:
     fixtures: Seq[Fixture],
     scalaVersion: ScalaVersion,
     backend: ScalaBackend
-  ): Seq[Dependency#WithVersion] = fixtures.map(_.framework.dependencyWithVersion(backend, scalaVersion, version = None))
+  ): Seq[Dependency#WithVersion] = fixtures.map(_.framework.dependencyWithVersion(
+    backend,
+    scalaLibrary = ScalaLibrary.fromScalaVersion(scalaVersion),
+    version = None
+  ))
   
   private def testTask(
     feature: Feature,